Add files via upload
Mini project “Calendar Application” is also a simple project built in C. This project has following features.  It displays a nicely formatted calendar of every month. You can find the day by entering the day, month and year. For example, if you enter day = 10, month = 03 and year = 1991, it gives you the day 'Sunday'. You can add the note for a particular day. If the given month has a note in it, it will display in that day.  You can play with the source code and modify as per your need